Added function to verify login
- Also added added the user information in Review
